change money location for later use in money-field
