+++ src/filters/date/money/money.sandbox.ts
@@ -4,8 +4,9 @@ import DropdownPlugin from '../../../components/dropdown/dropdown';
 import IntegerfieldPlugin from '../../../components/integerfield/integerfield';
 import { Enums } from '../../../utils/enums/enums';
 import { ENGLISH, FRENCH, Messages } from '../../../utils/i18n/i18n';
+import { MCurrencyType, MMoney, MMoneyFactory } from '../../../utils/money/money';
 import { ModulVue } from '../../../utils/vue/vue';
-import MoneyPlugin, { MCurrencyType, MMoney, MMoneyFactory } from './money';
+import MoneyPlugin from './money';
 import WithRender from './money.sandbox.html';
 
 @WithRender
